+++ CHANGELOG.md
@@ -15,6 +15,7 @@ Please avoid adding duplicate information across this changelog and JIRA/doc inp
 - RHPF-98: Log creation of API token. The token creation log message will trigger an administration event.
 - ROX-28716: New policy criterion "Days Since CVE Was Published" to allow creation of a policy that offers a grace period to teams to fix vulnerabilities within the number of days from when the CVE was published in the vulnerability feeds.
 - ROX-28296: Support for an OpenShift reencrypt route to expose Central (`central.exposure.route.reencrypt.enabled: true`).
+- ROX-28153: Support for Cosign keyless signing and verification of image signatures.
 
 ### Removed Features